requirements.txt
ib-insync>=0.9.86
redis>=4.5.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dateutil>=2.8.0
ta-lib>=0.4.25
matplotlib>=3.7.0
